Remove execute from non-pooled driver
